ProjMgtService-ImplementControllers: Implemented and defined methods for SprintRestController.
